requirements.txt
anyio==3.7.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.97.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.4.0
pytest-mock==3.11.1
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
